project/requirements.txt
affine==2.4.0
alembic==1.13.1
annotated-types==0.6.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
asyncpg==0.29.0
attrs==23.1.0
bcrypt==3.2.0
Bottleneck==1.3.7
branca==0.6.0
Brotli==1.0.9
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==2.0.4
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
comm==0.2.1
conda-merge==0.2.0
contextily==1.6.0
contourpy==1.2.0
cryptography==42.0.5
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
ecdsa==0.13
exceptiongroup==1.2.0
executing==0.8.3
fastapi==0.110.0
fiona==1.9.6
folium==0.14.0
fonttools==4.25.0
GDAL==3.8.5
geographiclib==2.0
geopandas==0.14.2
geopy==2.4.1
greenlet==3.0.1
h11==0.14.0
idna==3.4
importlib-metadata==7.0.1
importlib-resources==6.1.1
ipykernel==6.29.3
ipython==8.20.0
jedi==0.18.1
Jinja2==3.1.3
joblib==1.2.0
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.4
Mako==1.2.3
mapclassify==2.5.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mercantile==1.2.1
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
munkres==1.1.4
nest-asyncio==1.6.0
networkx==3.1
numexpr==2.8.7
numpy==1.26.4
osmnx==1.9.3
packaging==23.2
pandas==2.2.1
parso==0.8.3
passlib==1.7.4
pexpect==4.8.0
pillow==10.3.0
pip==23.3.1
platformdirs==3.10.0
prompt-toolkit==3.0.43
psutil==5.9.0
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pycparser==2.22
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.15.1
pyparsing==3.0.9
pyproj==3.6.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
rasterio==1.3.10
requests==2.31.0
rsa==4.9
Rtree==1.0.1
scikit-learn==1.3.0
scipy==1.12.0
setuptools==68.2.2
shapely==2.0.4
six==1.16.0
sniffio==1.3.0
snuggs==1.4.7
SQLAlchemy==2.0.25
sqlmodel==0.0.14
stack-data==0.2.0
starlette==0.36.3
threadpoolctl==2.2.0
tornado==6.3.3
traitlets==5.7.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.20.0
wcwidth==0.2.5
wheel==0.41.2
xyzservices==2022.9.0
zipp==3.17.0
